Createpols (#4)
* changes for new api and format

* removing

* comma

* fggggg

* deubg

* var change

* error

* act name

* wrong verb

* dfdfdfd

* syntax

* domain count

* domain count

* int

* int

* changes to count

* adding

* domaincount

* change to value

* cleanup unbind

* final cleanup

* removed output

* more cleanup

* adding ignore for counter

* file cleanup

* more options for config.  clean up code

* more

* typo

* fixes to test

* testing

* more tests

* error handling

* documentation around mynsconfig

+++ ns-copytons.py
@@ -3,13 +3,13 @@
 #USE AT OWN RISK
 
 #Imports
-import json, requests, base64, sys
+import json, requests, base64, sys, os
 #imports variables used for script
 from mynsconfig import *
 
 __author__ = "Ryan Butler (techdrabble.com)"
 __license__ = "GPL"
-__version__ = "1.0.0"
+__version__ = "2.0.0"
 __maintainer__ = "Ryan Butler"
 
 #what to perform
@@ -25,11 +25,13 @@ def getAuthCookie(connectiontype,nitroNSIP,nitroUser,nitroPass):
        }
    }
    payload = json.dumps(json_string)
-   response = requests.post(url, data=payload, headers=headers, verify=False)
-   response.raise_for_status()
+   try:
+     response = requests.post(url, data=payload, headers=headers, verify=False, timeout=1.0)
+   except requests.exceptions.RequestException as e:
+     sys.exit(1)
    cookie = response.cookies['NITRO_AUTH_TOKEN']
    nitroCookie = 'NITRO_AUTH_TOKEN=%s' % cookie
-   return nitroCookie
+   return nitroCookie 
 
 def logOut(connectiontype,nitroNSIP,authToken):
    url = '%s://%s/nitro/v1/config/logout' % (connectiontype, nitroNSIP)
@@ -65,35 +67,85 @@ def sendFile(connectiontype,nitroNSIP,authToken,nscert,localcert,nscertpath):
    }
    payload = json.dumps(json_string)
    response = requests.post(url, data=payload, headers=headers, verify=False)
-   print "CREATE CERT: %s" % response.reason
+   print "CREATING %s: %s" % (nscert, response.reason)
 
-def respPol(connectiontype,nitroNSIP,authToken,nsresppol,token_filename):
+def CreaterespPol(connectiontype,nitroNSIP,authToken,polname,token_filename,actname):
    url = '%s://%s/nitro/v1/config/responderpolicy' % (connectiontype, nitroNSIP)
    headers = {'Content-type': 'application/json','Cookie': authToken}
    buildrule = 'HTTP.REQ.URL.CONTAINS(\"well-known/acme-challenge/%s\")' % token_filename
-   print buildrule
    json_string = {
    "responderpolicy": {
-       "name": nsresppol,
+       "name": polname,
+       "action": actname,
        "rule": buildrule,}
    }
    payload = json.dumps(json_string)
-   response = requests.put(url, data=payload, headers=headers, verify=False)
-   print "EDIT RESPONDER POLICY: %s" % response.reason
+   response = requests.post(url, data=payload, headers=headers, verify=False)
+   print "CREATE RESPONDER POLICY: %s" % response.reason
 
-def respAct(connectiontype,nitroNSIP,authToken,nsrespact,token_value):
+def DeleterespPol(connectiontype,nitroNSIP,authToken,polname):
+   url = '%s://%s/nitro/v1/config/responderpolicy/%s' % (connectiontype, nitroNSIP, polname)
+   headers = {'Cookie': authToken}
+   response = requests.delete(url, headers=headers, verify=False)
+   print "DELETE RESPONDER POLICY: %s" % response.reason
+
+def DeleterespAct(connectiontype,nitroNSIP,authToken,actname):
+   url = '%s://%s/nitro/v1/config/responderaction/%s' % (connectiontype, nitroNSIP, actname)
+   headers = {'Cookie': authToken}
+   response = requests.delete(url, headers=headers, verify=False)
+   print "DELETE RESPONDER ACTION: %s" % response.reason
+
+def CreaterespAct(connectiontype,nitroNSIP,authToken,actname,token_value):
    url = '%s://%s/nitro/v1/config/responderaction' % (connectiontype, nitroNSIP)
    headers = {'Content-type': 'application/json','Cookie': authToken}
    buildtarget = "\"HTTP/1.0 200 OK\" +\"\\r\\n\\r\\n\" + \"%s\"" % token_value
-   print buildtarget
    json_string = {
    "responderaction": {
-       "name": nsrespact,
+       "name": actname,
+       "type": "respondwith",
        "target": buildtarget,}
    }
    payload = json.dumps(json_string)
+   response = requests.post(url, data=payload, headers=headers, verify=False)
+   print "CREATE RESPONDER ACTION: %s" % response.reason
+
+def BindrespPolCSW(connectiontype,nitroNSIP,authToken,polname,nsvip,domaincount):
+   url = '%s://%s/nitro/v1/config/csvserver_responderpolicy_binding' % (connectiontype, nitroNSIP)
+   headers = {'Content-type': 'application/json','Cookie': authToken}
+   json_string = {
+   "csvserver_responderpolicy_binding": {
+       "name": nsvip,
+       "policyname": polname,
+       "priority": domaincount,}
+   }
+   payload = json.dumps(json_string)
    response = requests.put(url, data=payload, headers=headers, verify=False)
-   print "EDIT RESPONDER POLICY: %s" % response.reason
+   print "BIND RESPONDER POLICY: %s" % response.reason
+
+def BindrespPolLB(connectiontype,nitroNSIP,authToken,polname,nsvip,domaincount):
+   url = '%s://%s/nitro/v1/config/csvserver_lbvserver_binding' % (connectiontype, nitroNSIP)
+   headers = {'Content-type': 'application/json','Cookie': authToken}
+   json_string = {
+   "csvserver_lbvserver_binding": {
+       "name": nsvip,
+       "policyname": polname,
+       "priority": domaincount,}
+   }
+   payload = json.dumps(json_string)
+   response = requests.put(url, data=payload, headers=headers, verify=False)
+   print "BIND RESPONDER POLICY: %s" % response.reason
+
+def UnBindrespPolCSW(connectiontype,nitroNSIP,authToken,polname,nsvip):
+   url = '%s://%s/nitro/v1/config/csvserver_responderpolicy_binding/%s?args=policyname:%s' % (connectiontype, nitroNSIP, nsvip, polname)
+   headers = {'Cookie': authToken}
+   response = requests.delete(url, headers=headers, verify=False)
+   print "UNBIND RESPONDER POLICY: %s" % response.reason
+
+def UnBindrespPolLB(connectiontype,nitroNSIP,authToken,polname,nsvip):
+   url = '%s://%s/nitro/v1/config/csvserver_lbvserver_binding/%s?args=policyname:%s' % (connectiontype, nitroNSIP, nsvip, polname)
+   headers = {'Cookie': authToken}
+   response = requests.delete(url, headers=headers, verify=False)
+   print "UNBIND RESPONDER POLICY: %s" % response.reason
 
 def removeFile(connectiontype,nitroNSIP,authToken,nscert,nscertpath):
    url = '%s://%s/nitro/v1/config/systemfile/%s?args=filelocation:%%2Fnsconfig%%2Fssl' % (connectiontype, nitroNSIP, nscert)
@@ -128,6 +180,12 @@ def createSSL(connectiontype,nitroNSIP,authToken, nscert, nspairname, nskey):
    response = requests.post(url, data=payload, headers=headers, verify=False)
    print "Create Netscaler CERT: %s" % response.reason
 
+def GetSSL(connectiontype,nitroNSIP,authToken, nspairname):
+   url = '%s://%s/nitro/v1/config/sslcertkey/%s' % (connectiontype, nitroNSIP, nspairname)
+   headers = {'Cookie': authToken}
+   response = requests.get(url,headers=headers, verify=False)
+   return response.status_code
+
 def createSSLCA(connectiontype,nitroNSIP,authToken,nscert,nspairname):
    url = '%s://%s/nitro/v1/config/sslcertkey' % (connectiontype, nitroNSIP)
    headers = {'Content-type': 'application/json','Cookie': authToken}
@@ -155,29 +213,57 @@ def linkSSL(connectiontype,nitroNSIP,authToken, nschainname, nspairname):
 authToken = getAuthCookie(connectiontype,nitroNSIP,nitroUser,nitroPass)
 if whattodo == "save":
    localcert = sys.argv[2]
-   print "Updating Netscaler Certificate"
-   removeFile(connectiontype,nitroNSIP,authToken,nscert,nscertpath)
-   sendFile(connectiontype,nitroNSIP,authToken,nscert,localcert,nscertpath)
-   updateSSL(connectiontype,nitroNSIP,authToken, nscert, nspairname)
+   localkey = sys.argv[3]
+   localchain = sys.argv[4]
+   existcode = GetSSL(connectiontype,nitroNSIP,authToken, nspairname)
+   if existcode == 200:
+       print "Using existing cert"
+       removeFile(connectiontype,nitroNSIP,authToken,nscert,nscertpath)
+       sendFile(connectiontype,nitroNSIP,authToken,nscert,localcert,nscertpath)
+       updateSSL(connectiontype,nitroNSIP,authToken, nscert, nspairname)
+   else:
+       print "Creating Netscaler Certificate"
+       sendFile(connectiontype,nitroNSIP,authToken,nscert,localcert,nscertpath)
+       sendFile(connectiontype,nitroNSIP,authToken,nskey,localkey,nscertpath)
+       sendFile(connectiontype,nitroNSIP,authToken,nschain,localchain,nscertpath)
+       createSSL(connectiontype,nitroNSIP,authToken, nscert, nspairname, nskey)
+       createSSLCA(connectiontype,nitroNSIP,authToken, nschain, nschainname)
+       linkSSL(connectiontype,nitroNSIP,authToken, nschainname, nspairname)
    SaveNSConfig(connectiontype,nitroNSIP,authToken)
 elif whattodo == "test":
    print "Connectivity To Netscaler OK"
 elif whattodo == "challenge":
-   print "Editing Challenge Policy"
    token_filename = sys.argv[2]
    token_value = sys.argv[3]
-   respPol(connectiontype,nitroNSIP,authToken,nsresppol,token_filename)
-   respAct(connectiontype,nitroNSIP,authToken,nsrespact,token_value)
-elif whattodo == "create":
-   localcert = sys.argv[2]
-   localkey = sys.argv[3]
-   localchain = sys.argv[4]
-   print "Create Netscaler Certificate"
-   sendFile(connectiontype,nitroNSIP,authToken,nscert,localcert,nscertpath)
-   sendFile(connectiontype,nitroNSIP,authToken,nskey,localkey,nscertpath)
-   sendFile(connectiontype,nitroNSIP,authToken,nschain,localchain,nscertpath)
-   createSSL(connectiontype,nitroNSIP,authToken, nscert, nspairname, nskey)
-   createSSLCA(connectiontype,nitroNSIP,authToken, nschain, nschainname)
-   linkSSL(connectiontype,nitroNSIP,authToken, nschainname, nspairname)
-   SaveNSConfig(connectiontype,nitroNSIP,authToken)
+   challenge_domain = sys.argv[4]
+   domaincount = int(sys.argv[5])
+   polname = '%s-%s' % (nsresppol, challenge_domain)
+   actname = '%s-%s' % (nsrespact, challenge_domain)
+   print "Creating Challenge Policy for %s" % challenge_domain
+   CreaterespAct(connectiontype,nitroNSIP,authToken,actname,token_value)
+   CreaterespPol(connectiontype,nitroNSIP,authToken,polname,token_filename,actname)
+   domaincount = polpristart + domaincount
+   
+   if viptype == "csw":
+       BindrespPolCSW(connectiontype,nitroNSIP,authToken,polname,nsvip,domaincount)
+   elif viptype == "lb":
+       BindrespPolLB(connectiontype,nitroNSIP,authToken,polname,nsvip)  
+   else:
+       print "Invalid VIP Type.  Check config"
+       os.exit()
+  
+elif whattodo == "clean":
+   challenge_domain = sys.argv[2]
+   polname = '%s-%s' % (nsresppol, challenge_domain)
+   actname = '%s-%s' % (nsrespact, challenge_domain)
+   print "Removing Challenge Policy for %s" % challenge_domain
+   if viptype == "csw":
+       UnBindrespPolCSW(connectiontype,nitroNSIP,authToken,polname,nsvip)
+   elif viptype == "lb":
+       UnBindrespPolLB(connectiontype,nitroNSIP,authToken,polname,nsvip)  
+   else:
+       sys.exit("Invalid VIP Type.  Check config")  
+   DeleterespPol(connectiontype,nitroNSIP,authToken,polname)
+   DeleterespAct(connectiontype,nitroNSIP,authToken,actname)
+   
 logOut(connectiontype,nitroNSIP,authToken)
